Fx impl, fixes

+++ src/main/kotlin/drawing/DrawingApi.kt
@@ -5,6 +5,6 @@ data class Coordinate(val x: Double, val y: Double)
 interface DrawingApi {
     fun getDrawingAreaWidth(): Double
     fun getDrawingAreaHeight(): Double
-    fun drawCircle(leftTop: Coordinate, radius: Int)
+    fun drawCircle(leftTop: Coordinate, radius: Double)
     fun drawLine(begin: Coordinate, end: Coordinate)
 }
